+++ core/basic/AdminBar.php
@@ -26,7 +26,24 @@ class AdminBar {
 
 	public function style() {
 		?>
-        <style>@media screen and (max-width: 782px) {
+        <style>
+            #wpadminbar .coreactivity-adminbar-count {
+                display: inline-block;
+                text-align: center;
+                padding: 0 6px;
+                height: 20px;
+                border-radius: 10px;
+                margin: 6px 0 0 4px;
+                vertical-align: top;
+                line-height: 20px;
+                font-size: 12px
+            }
+
+            #wpadminbar .ab-submenu .coreactivity-adminbar-count {
+                float: right;
+            }
+
+            @media screen and (max-width: 782px) {
                 #wpadminbar li#wp-admin-bar-coreactivity-menu {
                     display: block;
                 }
@@ -38,31 +55,58 @@ class AdminBar {
 		if ( current_user_can( 'manage_options' ) ) {
 			global $wp_admin_bar;
 
+			$show  = '';
+			$count = coreactivity_settings()->get( 'admin_bar_indicator' ) ? DB::instance()->get_new_log_entries_since_last_log_visit() : 0;
+
+			$title = '<span style="margin-top: 2px" class="ab-icon dashicons dashicons-database"></span><span class="ab-label">' . __( 'coreActivity', 'coreactivity' ) . '</span>';
+
+			if ( $count > 0 ) {
+				$show  = '<span class="wp-ui-notification coreactivity-adminbar-count">' . $count . '</span>';
+				$title .= $show;
+			}
+
 			$wp_admin_bar->add_menu( array(
 				'id'    => 'coreactivity-menu',
-				'title' => '<span style="margin-top: 2px" class="ab-icon dashicons dashicons-database"></span><span class="ab-label">' . __( 'coreActivity', 'coreactivity' ) . '</span>',
+				'title' => $title,
 				'href'  => network_admin_url( 'admin.php?page=coreactivity-dashboard' ),
 			) );
 
-			$wp_admin_bar->add_menu( array(
+			$wp_admin_bar->add_group( array(
 				'parent' => 'coreactivity-menu',
-				'id'     => 'coreactivity-menu-logs',
-				'title'  => __( 'Logs', 'coreactivity' ),
-				'href'   => network_admin_url( 'admin.php?page=coreactivity-logs' ),
+				'id'     => 'coreactivity-menu-top',
+			) );
+
+			$wp_admin_bar->add_group( array(
+				'parent' => 'coreactivity-menu',
+				'id'     => 'coreactivity-menu-bottom',
 			) );
 
 			$wp_admin_bar->add_menu( array(
-				'parent' => 'coreactivity-menu',
+				'parent' => 'coreactivity-menu-top',
 				'id'     => 'coreactivity-menu-events',
 				'title'  => __( 'Events', 'coreactivity' ),
 				'href'   => network_admin_url( 'admin.php?page=coreactivity-events' ),
 			) );
 
 			$wp_admin_bar->add_menu( array(
-				'parent' => 'coreactivity-menu',
+				'parent' => 'coreactivity-menu-top',
+				'id'     => 'coreactivity-menu-logs',
+				'title'  => __( 'Logs', 'coreactivity' ) . $show,
+				'href'   => network_admin_url( 'admin.php?page=coreactivity-logs' ),
+			) );
+
+			$wp_admin_bar->add_menu( array(
+				'parent' => 'coreactivity-menu-bottom',
 				'id'     => 'coreactivity-menu-settings',
 				'title'  => __( 'Settings', 'coreactivity' ),
 				'href'   => network_admin_url( 'admin.php?page=coreactivity-settings' ),
+			) );
+
+			$wp_admin_bar->add_menu( array(
+				'parent' => 'coreactivity-menu-bottom',
+				'id'     => 'coreactivity-menu-tools',
+				'title'  => __( 'Tools', 'coreactivity' ),
+				'href'   => network_admin_url( 'admin.php?page=coreactivity-tools' ),
 			) );
 		}
 	}