+++ core/log/Upgrader.php
@@ -2,6 +2,9 @@
 
 namespace Dev4Press\Plugin\CoreActivity\Log;
 
+use WP_Error;
+use WP_Theme;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
@@ -33,7 +36,7 @@ class Upgrader {
 			'plugins_updates' => $plugins->response ?? array(),
 		);
 
-		update_site_option( 'coresecurity_temp_plugins_themes', $backup, false );
+		update_site_option( 'coreactivity_temp_plugins_themes', $backup, false );
 
 		return $value;
 	}
@@ -42,6 +45,11 @@ class Upgrader {
 		if ( isset( $data['type'] ) && isset( $data['action'] ) ) {
 			$_type   = $data['type'];
 			$_action = $data['action'];
+			$_error  = false;
+
+			if ( $obj->skin->result instanceof WP_Error ) {
+				$_error = $obj->skin->result;
+			}
 
 			if ( 'plugin' == $_type && 'update' == $_action ) {
 				$plugins = isset( $data['plugins'] ) ? (array) $data['plugins'] : array();
@@ -51,27 +59,79 @@ class Upgrader {
 					$previous = $this->get_plugin_previous_version( $plugin_code );
 					$package  = $this->get_plugin_package_url( $plugin_code );
 
-					do_action( 'coreactivity_upgrader_plugin_update', $plugin_code, $plugin, $previous, $package );
+					if ( $_error ) {
+						do_action( 'coreactivity_upgrader_plugin_update_error', $plugin_code, $plugin, $previous, $package, $_error );
+					} else {
+						do_action( 'coreactivity_upgrader_plugin_update', $plugin_code, $plugin, $previous, $package );
+					}
 				}
 			} else if ( 'plugin' == $_type && 'install' == $_action ) {
+				$plugin      = $obj->new_plugin_data ?? array();
+				$plugin_code = $obj->plugin_info();
 
+				if ( $_error ) {
+					do_action( 'coreactivity_upgrader_plugin_install_error', $plugin_code, $plugin, $_error );
+				} else {
+					do_action( 'coreactivity_upgrader_plugin_install', $plugin_code, $plugin );
+				}
+			} else if ( 'theme' == $_type && 'update' == $_action ) {
+				$themes = isset( $data['themes'] ) ? (array) $data['themes'] : array();
+
+				foreach ( $themes as $theme_code ) {
+					$theme    = wp_get_theme( $theme_code );
+					$previous = $this->get_theme_previous_version( $theme_code );
+					$package  = $this->get_theme_package_url( $theme_code );
+
+					if ( $theme instanceof WP_Theme ) {
+						if ( $_error ) {
+							do_action( 'coreactivity_upgrader_theme_update_error', $theme_code, $theme, $previous, $package, $_error );
+						} else {
+							do_action( 'coreactivity_upgrader_theme_update', $theme_code, $theme, $previous, $package );
+						}
+					}
+				}
+			} else if ( 'theme' == $_type && 'install' == $_action ) {
+				$theme      = $obj->new_theme_data ?? array();
+				$theme_code = $obj->result['destination_name'] ?? '';
+
+				if ( $_error ) {
+					do_action( 'coreactivity_upgrader_theme_install_error', $theme_code, $theme, $_error );
+				} else {
+					do_action( 'coreactivity_upgrader_theme_install', $theme_code, $theme );
+				}
 			}
 		}
 
-		delete_site_option( 'coresecurity_temp_plugins_themes' );
+		delete_site_option( 'coreactivity_temp_plugins_themes' );
 	}
 
 	protected function get_plugin_package_url( $plugin_code ) {
-		$data   = get_site_option( 'coresecurity_temp_plugins_themes', array() );
+		$data   = get_site_option( 'coreactivity_temp_plugins_themes', array() );
 		$plugin = $data['plugins_updates'][ $plugin_code ] ?? array();
+		$plugin = (array) $plugin;
 
-		return empty( $plugin ) ? '' : ( $plugin->package ?? '' );
+		return empty( $plugin ) ? '' : ( $plugin['package'] ?? '' );
+	}
+
+	protected function get_theme_package_url( $theme_code ) {
+		$data  = get_site_option( 'coreactivity_temp_plugins_themes', array() );
+		$theme = $data['themes_updates'][ $theme_code ] ?? array();
+		$theme = (array) $theme;
+
+		return empty( $theme ) ? '' : ( $theme['package'] ?? '' );
 	}
 
 	protected function get_plugin_previous_version( $plugin_code ) {
-		$data   = get_site_option( 'coresecurity_temp_plugins_themes', array() );
+		$data   = get_site_option( 'coreactivity_temp_plugins_themes', array() );
 		$plugin = $data['plugins'][ $plugin_code ] ?? array();
 
 		return $plugin['Version'] ?? '';
+	}
+
+	protected function get_theme_previous_version( $theme_code ) {
+		$data  = get_site_option( 'coreactivity_temp_plugins_themes', array() );
+		$theme = $data['themes'][ $theme_code ] ?? array();
+
+		return $theme['Version'] ?? '';
 	}
 }