+++ core/plugins/DebugPress.php
@@ -14,6 +14,7 @@ class DebugPress extends Plugin {
 	protected $icon = 'plugin-debugpress';
 	protected $object_type = 'phperror';
 	protected $plugin_file = 'debugpress/debugpress.php';
+	protected $skip_duplicates = true;
 
 	public function init() {
 		add_filter( 'coreactivity_debugpress_ajax_call_log_active', array( $this, 'skip_heartbeat' ), 10, 3 );